==> 1_EDA/1. Location EDA.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

##load in datasets

#1. test.csv
test = pd.read_csv('./datasets/test.csv')
test = pd.DataFrame(test)
test.head()


# Don't need to run again
##load in datasets

#1. train.csv
#wn = pd.read_csv('./datasets/train.csv')
#wn = pd.DataFrame(train)
#wn.head()

test.isnull().sum() #Checking for missing data

test.info() #Checking the info of the dataframe

test.shape #Checking size 

test.columns # Looking at columns 

### NumMosquitos is missing here but is in the training data 

wn = pd.read_csv('./datasets/train.csv')

wn.shape #Train shape

wn.columns #Train columns

# test.columns

wn_features = ['Date','Species', 'Trap', 'Longitude', 'Latitude'] #checking out the features 
target = ['WnvPresent']

wn[wn_features].isna().sum() # checking for nan values

test[wn_features].isna().sum() #checking for nan values 

wn[wn_features].info()

test[wn_features].info()

wn.shape  #dropped duplicates 

wn['WnvPresent'].value_counts(normalize=True) #still unbalanced classes

west_nile_features_train = wn[wn_features]

west_nile_features_test = test[wn_features]

west_nile_features_train.to_csv('clean_train.csv', index=False)
west_nile_features_test.to_csv('clean_test.csv', index=False)

west_nile_features_train.head()

west_nile_features_test.head()

west_nile_features_test.shape

west_nile_features_train.shape

target = pd.DataFrame(wn['WnvPresent'])
target.to_csv('target.csv', index=False)
