==> 1_EDA/2. Train Data EDA.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

##load in datasets

#1. train.csv
train = pd.read_csv('./datasets/train.csv')
train = pd.DataFrame(train)
train.head()

train.isnull().sum()

train.columns

train.info()

train.describe()

train.shape

train['Trap']
