==> 1_EDA/4. weather_dataset2.ipynb <==
import pandas as pd
import numpy as np

#load datasets
train = pd.DataFrame(pd.read_csv('../datasets/train.csv'))
test = pd.DataFrame(pd.read_csv('../datasets/test.csv'))
spray = pd.DataFrame(pd.read_csv('../datasets/spray.csv'))
weather = pd.DataFrame(pd.read_csv('../datasets/weather.csv'))

#8. create weather_new df with index as Station
weather_new = weather.set_index('Station')

#9. apply average logic from step 7 to weather_new df
    #for both cols with M values, imputing 0.0
        #mostly affects col Water1
newcols = {}

for col in weather_new.drop(['Date', 'CodeSum'], axis = 1):
    newvals = []
    
    for date in weather_new['Date'].unique():
        val1 = weather_new[weather_new['Date'] == date].loc[1, col]
        val2 = weather_new[weather_new['Date'] == date].loc[2, col]
        
        if val1 not in ['M', '-', ' ', '  T'] and val2 not in ['M', '-', ' ', '  T']:
            aveval = np.average([pd.to_numeric(val1), pd.to_numeric(val2)])
            newvals.append(aveval)
            #print (col, date, aveval)
        elif val1 not in ['M', '-', ' ', '  T'] and val2 in ['M', '-', ' ', '  T']:
            newvals.append(val1)
            #print (col, date, val1)
        elif val1 in ['M', '-', ' ', '  T'] and val2 not in ['M', '-', ' ', '  T']:
            newvals.append(val2)
            #print (col, date, val2)
        else:
            newvals.append(0.0)
            #print ('Problem')
            
    newcols[col] = newvals
    #print (newvals)
print (newcols.keys())

#10. create codesum df for handling CodeSum text abbreviations (denotes weather phenomena like thunderstorm, snow, smoke, etc)
codesum1 = weather[weather['Station'] == 1]['CodeSum'].to_frame()
codesum1 = codesum1.set_index(weather[weather['Station'] == 1]['Date'])
codesum1.index.rename('Date')

codesum2 = weather[weather['Station'] == 2]['CodeSum'].to_frame()
codesum2 = codesum2.set_index(weather[weather['Station'] == 2]['Date'])
codesum2.index.rename('Date')

codesum = codesum1.join(codesum2, lsuffix = '1', rsuffix = '2')
print(codesum.shape)

#11. check if CodeSum from each Station matches
    #a. if match: use CodeSum
    #b. if no_match: use CodeSum with more information (non-empty Codesum, aggregate Codesum)

exact_matches = 0
both_null = 0
partial_matches = 0
stat1_null = 0
stat2_null = 0

aggregate = []

for date in codesum.index:
    #print(date)
    
    if codesum['CodeSum1'][date] == ' ' and codesum['CodeSum2'][date] == ' ':
        both_null += 1
        aggregate.append(' ')
    elif codesum['CodeSum1'][date] == codesum['CodeSum2'][date]:
        exact_matches += 1
        aggregate.append(codesum['CodeSum1'][date])
        #print(matches)
    else:
        
        if len(codesum['CodeSum1'][date]) == 1 and len(codesum['CodeSum2'][date]) > 1:
            stat1_null +=1
            aggregate.append(codesum['CodeSum2'][date])
            #print(date, codesum['CodeSum1'][date], ',', codesum['CodeSum2'][date])
        elif len(codesum['CodeSum1'][date]) > 1 and len(codesum['CodeSum2'][date]) == 1:
            stat2_null += 1
            aggregate.append(codesum['CodeSum1'][date])
            #print(date, codesum['CodeSum1'][date], ',', codesum['CodeSum2'][date])
        else:
            partial_matches += 1
            agg = codesum['CodeSum1'][date].split() + codesum['CodeSum2'][date].split()
            agg = pd.Series(list(set(agg)))
            agg = agg.str.cat(sep = ' ')
            aggregate.append(agg)
            #print(date, codesum['CodeSum1'][date], ',', codesum['CodeSum2'][date])            
            #print(agg)
            #print ('\n')    
        #print (no_matches)

print('\n',
      'exact_matches:', exact_matches, '\n',
      'both_null:', both_null, '\n',
      'partial_matches:', partial_matches, '\n',
      'stat1_null:', stat1_null, '\n',
      'stat2_null:', stat2_null, '\n',
      '\n',
      'total: ', exact_matches + both_null + stat1_null + stat2_null + partial_matches)

aggregate = pd.Series(aggregate)

#11b. checker
agg = codesum['CodeSum1']['2007-05-31'].split() + codesum['CodeSum2']['2007-05-31'].split()
print(codesum['CodeSum1']['2007-05-31'], ',', codesum['CodeSum2']['2007-05-31'])

agg = pd.Series(list(set(agg)))
agg = agg.str.cat(sep = ' ')
agg

#12. update weather_new df
    #a.Repeat step #8 first: create weather_new df with index as Station
    #b. set unique dates
    #c. set calculated values

weather_new = weather.set_index('Station')
#weather_new.head()

#12b. set unique dates
weather_new = weather_new.loc[[1], :]
weather_new = weather_new.reset_index().drop(['Station'], axis = 1)
#weather_new.head()

#12c. set calculated values
for col in weather_new.drop(['Date', 'CodeSum'], axis = 1).columns:
    weather_new[col] = [newcols[col][i] for i in weather_new[col].index]

weather_new['CodeSum'] = [aggregate[i] for i in weather_new['CodeSum'].index]
#weather_new['CodeSum'][0:10]

#13. checking against original weather df
weather_new.head(10)
#weather.head(20)

#14. Split weather_new df into train / test
    #train_set = years 2007, 09, 11, 13
    #test_set = years 2008, 10, 12, 14
    #spray = years 2011, 13
    #weather = years 2007-2014

#year in weather_new df
weather_new['Year'] = [int(weather_new['Date'][i][0:4]) for i in weather_new['Date'].index]

train_yrs = [2007, 2009, 2011, 2013]
test_yrs = [2008, 2010, 2012, 2014]

train_weather = weather_new.loc[weather_new['Year'].isin(train_yrs)]
test_weather = weather_new.loc[weather_new['Year'].isin(test_yrs)]

#print('train:', train_weather.shape, 'test:', test_weather.shape)
#print(weather_new[weather_new['Year'] == 2007].shape)
#train_weather.head(10)

test_weather.head(5)

#15. Handling merging training dfs on with different dates
    #len(list(pd.unique(test['Date']))) returns 95

#option 1: between last spray date and the date of spray, what is the average weather data?
    #for date == 2007-05-29, last 7 days worth of weather values --> prev_days = 7
    #for date == 2007-06-05, days between 2007-05-29 and 2007-06-05 worth of weather values --> prev_days = 7
    #for date == 2007-06-26, days between 2007-06-5 and 2007-06-26 worth of weather values --> prev_days = 21  
#caveat: interval of days is not be consistent

#option 2: straight 7 days prior worth of weather data
    #for any date, last 7 days worth of weather values --> prev days = 7

#option 3: month of date
    #for any date, average of entire month's worth of weather values

#option 4: day of date
    #for any date, weather values on matched date

#15. Option 4: combining datasets
    #spray df has no data past 2011/2013
combined = test.merge(test_weather, on = 'Date')
#combined.head()

#15. Option 2: for train_weather values between test['Date'] and test['Date'] - 7, average_values for test_weather[col]

#15-1. create new column in train_weather that is test['Date']
test_weather['Dateloc'] = pd.to_datetime(test['Date'])

    #convert date columns in train_weather to datetime type
test_weather['Date'] = pd.to_datetime(test_weather['Date'])

#15-2. create new column in train_weather that is train['Date'] - 7
from datetime import timedelta
test_weather['Prev7date'] = test_weather['Dateloc'] - timedelta(days = 7)

#15-3. create lookup table df
lookup = test_weather.iloc[:, :].drop(['Year', 'Dateloc', 'Prev7date'], axis = 1)
#lookup.set_index(['Date'], inplace = True)
lookup.head()

test_weather.head()

lookup.info()

#15-4. update test_weather df: drop unnecessary rows
    #keep unique dates from test set (95 unique values)

test_weather = test_weather.reset_index()
test_weather = test_weather.drop(['index'], axis = 1)

test_weather = test_weather.loc[:94, :]

test_weather['Dateloc'] = list(test['Date'].unique())
test_weather['Prev7date'] = pd.to_datetime(test_weather['Dateloc']) - timedelta(days = 7)
test_weather = test_weather.drop(['Date'], axis = 1)

#test_weather.head()

#15-5a. numerical cols: calculate prev7day_average values by unique test['Date']

allcols = {}

for col in test_weather.drop(['CodeSum', 'Year', 'Dateloc', 'Prev7date'], axis = 1).columns:
    #print(col)
    colvals = []
    
    for date in test_weather['Dateloc']:
        vals = []
        t = 7
    
        while t > 0:
            lookup_date = pd.to_datetime(date) - timedelta(days = t)
            val = lookup[lookup['Date'] == lookup_date][col].values[0]
            vals.append(val)
            #print(t, lookup_date, val)
            t -= 1
        aveval = np.average(pd.to_numeric(vals)).round(1)
        #print(vals)
        #print('dateloc:', date, aveval, '\n')
        
        colvals.append(aveval)
    
    #print(colvals)
    allcols[col] = colvals

print(allcols.keys())

#15-5b. text cols: aggregate all values by unique test['Date']
    #col CodeSum

for col in test_weather.loc[:, ['CodeSum']].columns:
    #print(col)
    colvals = {}
    
    for date in test_weather['Dateloc']:
        #print(date)
        vals = []
        t = 7
    
        while t > 0:
            lookup_date = pd.to_datetime(date) - timedelta(days = t)
            val = lookup[lookup['Date'] == lookup_date][col].values[0].split()
            vals.append(val)
            #print(t, lookup_date, val)
            t -= 1

        #flatten list of lists into one list + return unique CodeSum values
        vals = [item for sublist in vals for item in sublist]
        vals = list(set(vals))
        #print(date, vals)
        
        colvals[date] = vals
        #print('\n')
    
    #print(colvals.items())


#15-5. checker
#print(allcols.items())
#allcols['Tmax'][0:10]
#print(colvals.items())

#15-6. input prev7day_ave values into train_weather by col
    #a. numeric cols
    #b. text col
for col in test_weather.drop(['Year', 'Dateloc', 'Prev7date', 'CodeSum'], axis = 1).columns:
    test_weather[col] = [allcols[col][i] for i in test_weather[col].index]

codesums = [colvals[date] for date in test_weather['Dateloc']]
test_weather['CodeSum'] = [codesums[row] for row in range(len(test_weather['CodeSum']))]

#test_weather.head()
#test_weather.loc[:, ['Dateloc', 'CodeSum']].head()

#16. clean train_weather: drop unnecessary cols, rename Dateloc to Date
test_weather = test_weather.drop(['Year', 'Prev7date'], axis = 1)
test_weather.rename(columns={'Dateloc': 'Date'}, inplace=True)
#test_weather.head()

#17. merge train loc + weather dfs together
    #spray has different dates
test_locwea = test.merge(test_weather, on = 'Date')

#test_locwea.head()
#test_locwea.columns

#18. save test_weather to_csv
test_weather.to_csv('test_weather')

#making some changes to test git fetch
