==> 1_EDA/5. weather_dataset3.ipynb <==
import pandas as pd
#creating a tableau view of weather stations + sprayed areas

spray = pd.read_csv('../datasets/spray.csv')
spray.info()

weatherstats = pd.DataFrame([['2011-01-01', '12:00:00 AM', 41.995, -87.933], ['2011-01-01', '12:00:00 AM', 41.786, -87.752]], columns=['Date', 'Time', 'Latitude', 'Longitude'])

spray = spray.append(weatherstats)
spray

spray.to_csv('spray_weather.csv')
