==> 1_EDA/6. merge.ipynb <==
import pandas as pd


train_loc = pd.read_csv('../datasets/clean_train.csv')
train_wea = pd.read_csv('../datasets/train_weather')

train_merge_w = train_loc.merge(train_wea, on = 'Date')

test_loc = pd.read_csv('../datasets/clean_test.csv')
test_wea = pd.read_csv('../datasets/test_weather')

test_merge_w = test_loc.merge(test_wea, on = 'Date')

train_merge_w.to_csv('../datasets/train_merge_w.csv')
test_merge_w.to_csv('../datasets/test_merge_w.csv')
