==> 1_EDA/7. spray eda.ipynb <==
import pandas as pd

%matplotlib inline

# load spray data

spray = pd.read_csv('.../datasets/spray.csv')

# investigate spray data

spray.shape

spray.head()

spray.info()

## need to change type of date and time to datetime (keeping pure date column for grouping)

spray['Time'] = pd.to_datetime(spray['Date']+' '+spray['Time'])
spray['Date'] = pd.to_datetime(spray['Date'])

spray.head()

spray.describe()

 ## verified with Google maps that these are reasonable values

spray['Date'].value_counts().sort_index()

## only ten dates when pesticides were sprayed

spray.drop_duplicates().shape

## this shows that there was more than one spray in one location on a single day

# look at spray summary
spray.groupby('Date').agg(['min','max'])

# save to csv
spray.to_csv('./assets/spray_clean.csv')
